+++ rust/target/wasm32-unknown-unknown/release/libhexgrid_wasm.d
@@ -1,0 +1,1 @@
+/Users/buley/Documents/Code/hexgrid-3d/rust/target/wasm32-unknown-unknown/release/libhexgrid_wasm.rlib: /Users/buley/Documents/Code/hexgrid-3d/rust/src/lib.rs /Users/buley/Documents/Code/hexgrid-3d/rust/src/math.rs /Users/buley/Documents/Code/hexgrid-3d/rust/src/spatial.rs /Users/buley/Documents/Code/hexgrid-3d/rust/src/statistics.rs